requirements.txt
attrs==23.1.0
beautifulsoup4==4.12.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
et-xmlfile==1.1.0
exceptiongroup==1.1.2
google-api-core==2.11.1
google-api-python-client==2.95.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.1
h11==0.14.0
httplib2==0.22.0
idna==3.4
lxml==4.9.3
mysql-connector-python==8.1.0
numpy==1.25.1
oauthlib==3.2.2
openpyxl==3.1.2
outcome==1.2.0
pandas==2.0.3
protobuf==4.21.12
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.1.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
selenium==4.10.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
tqdm==4.65.0
trio==0.22.2
trio-websocket==0.10.3
tzdata==2023.3
undetected-chromedriver==3.5.0
uritemplate==4.1.1
urllib3==1.26.16
websockets==11.0.3
wsproto==1.2.0
